[ISSUE-27] Refactored NavHeader component + adjustments
