--- src/Components/Pages/Landing/footer/data.js
@@ -1,5 +1,0 @@
-export const data = [
-  ["app name", "home", "beta", "sponsors"],
-  ["legal stuff", "terms", "privacy", "security"],
-  ["company", "about", "press", "jobs"]
-];
