--- src/Components/Pages/Landing/slider/data.js
@@ -1,25 +1,0 @@
-export const data = [
-  {":dizzy:": "512"},
-  {":fire:": "10.2k"},
-  {":raising_hand:": "13.4k"},
-  {":grapes:": "390"},
-  {":cold_sweat:": "220.8k"},
-  {":dizzy:": "512"},
-  {":fire:": "10.2k"},
-  {":raising_hand:": "13.4k"},
-  {":grapes:": "390"},
-  {":cold_sweat:": "220.8k"}
-];
-
-export const images = [
-  "/jj-jordan-1365547-unsplash.jpg",
-  "/ofido-henry-1366004-unsplash.jpeg",
-  "/ana-francisconi-1364352-unsplash.jpg",
-  "/mihai-stefan-1360937-unsplash.jpg",
-  "/parker-whitson-1362886-unsplash.jpg",
-  "/jj-jordan-1365547-unsplash.jpg",
-  "/ofido-henry-1366004-unsplash.jpeg",
-  "/ana-francisconi-1364352-unsplash.jpg",
-  "/mihai-stefan-1360937-unsplash.jpg",
-  "/parker-whitson-1362886-unsplash.jpg"
-];
